\name{CC164}
\alias{CC164}
\title{Cochran and Cox, p164}
\description{
 Factorial Experiment
}
\usage{
	data(CC164)
}

\details{

	A three factor experiment involving lettuce plants. Response is
	the number of lettuce plants emerging after treatment by three
	levels of either nitrogen or phospate fertilizer.
}
\references{
	Chochran, W. and Cox, G. (1957). Experimental Design, 2nd Ed.
		John Wiley & Sons, New York. 
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

