\name{composite}
\alias{composite}
\title{Test of theroplastic composite}
\description{
  Data from a factorial experiment.
}
\usage{
	data(composite)
}

\details{

	A two variable factorial experiment to test the strength of a thermoplastic
	composite to laser intensity and tape speed. 
}
\references{
	Faraway, J. (2005). Linear Models with R. Chapman and Hall. N.Y. 
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

