\name{Federer276Numeric}
\alias{Federer276Numeric}
\title{Federer 276 Numeric}
\description{
 Randomized block in three replicates
}
\usage{
	data(Federer276)
}

\details{

	Varieties and Treatments are numeric variables.

	A seed-gemination test on eight varieties of guayule. Four treatments 
	were applied to lots of seed from the varieties. 

	A split plot design assigning varieties to whole plots was used. Treatments
	formed randomized blocks within whole plots. There were three replicates.

	The whole plots were nested within replicates
}
\references{
	Federer, W.T. (1955). Experimental Design, Macmillan, N.Y. 
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

