\name{ghoctane}
\alias{ghoctane}
\title{Octane blends}
\description{
 Gasoline blending mixture experiment
}
\usage{
	data(ghoctane)
}

\details{

	An octane-blending experiment for a ternary system. Motor-octane ratings
	were obtained for a seven lattice point design, and two checkpoints.

}
\references{
	Gorman, J.W. and Hinman, J.E.  (1962). Simples lattice designs for
		multicomponent systems. Technometics. 4-4. 463-487.
		John Wiley & Sons, New York. 
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

