\name{Hald17.4}
\alias{Hald17.4}
\title{Hald Data}
\description{
 Data from Hald Table 17.4 on p 505.
}
\usage{
	data(Hald17.4)
}

\details{

	A randomized block experiment with 7 blocks and 6 treatments.
}
\references{
	Hald, A. (1952). Statistical theory with engineering applications. Wiley. NY.
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

