\name{manly126}
\alias{manly126}
\title{Manly, p126}
\description{
 Lizard data
}
\usage{
	data(manly126)
}

\details{

	Two factor data showing ant consumption by lizards. In addition,
	three aditional columns, v1,v2,v3, are included representing
	the additions used by Manly to calculate power.
}
\references{
	Manly, B.F.J. (1998). Randomization, bootstrap and Monte Carlo
		methods in Biology. 2nd Ed. Chapman & Hall, London.
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

