\name{manly136}
\alias{manly136}
\title{Non normal Ephemeroptera counts}
\description{
	Ephemeroptera counts from New Zealand streams.
}
\usage{
	data(manly136)
}

\references{
	Manly, B.F. (1998). Randomization, bootstrap and monte carlo methods in biology.
		Chapman and Hall, NY.
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

