\name{NPK}
\alias{NPK}
\title{NPK Data}
\description{
 Data from Venables and Ripley, Table 6.1, p176.
}
\usage{
	data(NPK)
}

\details{

	A fractional factorial with blocks confounded with the three way interaction.

}
\references{
	Venables, W.N. and Ripley, B.D. (2000) Modern applied statistics with S-Plus. 3rd ed. Springer, NY.
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

