\name{Plasma}
\alias{Plasma}
\title{Plasma Etching}
\description{
	An experiment designed to guide optimization of a nitride etch process
}
\usage{
	data(Plasma)
}

\details{

	An orthogonal arry, capable of supporting a linear model only. The 
	factors are denoted by their units of measurement. They were W, 
	the power applied to the cathode; mTorr, the pressure
	in the reaction chamber; cm, gap between anode and cathode; sscm, flow
	if the reactant gas. The three responses: Amin, etch rate; Pct, unifomity;
	sinpoly, selectivity. 
}
\references{
	Vardeman, Stephen B. (1994).Statistics for engineering problem solving.
		PWS publishing Co. Boston. p 596
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

