\name{Quinlan}
\alias{Quinlan}
\title{Quinlan's data}
\description{
 Data from a Symposium on Taguchi Methods, analyzed by G.E.P. Box
}
\usage{
	data(Quinlan)
}

\details{

	A saturated design: L16 in Taguchi termonology, but just a 2^(15-11)III design using the symbolism in Box (1978).
	Reanalyzed by Box (1988).
}
\references{
	\describe{
	\item{Quinlan, J. (1985)}{"Product improvement by application of Taguchi methods." in American Supplier Institute News
		(special symposium ed.) Dearborn, MI. American Supplier Institute. 11-16.}
	\item{Box, G.E.P., Hunter, W.G. and Hunter J.S. (1978)}{ Statistics for Experimenters. Wiley, N.Y}
	\item{Box, G. (1988)}{Signal-to-noise ratios, performance criteria, and transformations. Technometics. 30-1. 1-17.}
	}
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

