\name{ratGenotype}
\alias{ratGenotype}
\title{ratGenotype Data}
\description{
 Data from Scheffe, p140
}
\usage{
	data(ratGenotype)
}

\details{
	An unbalanced design. The data originally appeared in Bailey (1953), and was reproduced in Scheffe (1959)
	with the description:

	\dQuote{...gives the weight of hybrid female rats in a foster-nursing experiment with four types of
	rats. (The weights are litter averges in grams at 28 days. The within litter variance was obviously
	negligible compared to the between-litter variance.) The factors in the two-way laout are the genotype
	of the foster mother and that of the litter.
	}
}
\references{
	\describe{
	\item{Scheffe, H. (1959)}{The Analysis of Variance, Wiley, NY.}
	\item{Bailey, D.W. (1953)}{``The inheritance of Maternal Influences on the Growth of the Rat''.
		Ph.D. thesis, Univ. California}
	}
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

