\name{wool}
\alias{wool}
\title{wool data}
\description{
  A three factor experiment with non-normal response
}
\usage{
	data(wool)
}

\details{

	There are three, 3-level factors: 
	\describe{
	\item{len}{length of test specimen in mm}
	\item{amp}{amplitude of loading cycle in mm}
	\item{load}{the load in g}
	}
	The response was the number of cycles to failure of worsted yarn under cycles of repeated loading

}
\references{
	Box, G.E.P. and Cox, D.R. (1964) An analysis of transformations (with discussion). J.R. Statists. Soc. B. 26. 211-246.
}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

